refactor(Attacker): Attacker의 구현체들을 Attacker와 Piece을 상속받도록 변경
